+++ src/Crypto/WebAuthn/AttestationStatementFormat/AndroidSafetyNet.hs
@@ -121,16 +121,25 @@ instance Aeson.ToJSON Statement where
 data VerificationError
   = -- | The receiced nonce was not set to the concatenation of the
     -- authenticator data and client data hash
-    VerificationErrorInvalidNonce
-  | -- | The response was created to far in the past
-    -- (first: now, second: generated time)
-    VerificationErrorResponseTooOld HG.DateTime HG.DateTime
-  | -- | The response was created to far in the future
-    -- (first: now, second: generated time)
-    VerificationErrorResponseInFuture HG.DateTime HG.DateTime
+    NonceMismatch
+      { -- | Nonce from the AndroidSafetyNet response
+        responseNonce :: Text,
+        -- | Base64 encoding of the SHA-256 hash of the concatenation of
+        -- authenticatorData and clientDataHash
+        calculatedNonce :: Text
+      }
+  | -- | The response was created to far in the past or future
+    ResponseTimeInvalid
+      { -- | The UTC time minus the allowed drift specified in the `Format`.
+        lowerBound :: HG.DateTime,
+        -- | The UTC time plus the allowed drift specified in the `Format`.
+        upperBound :: HG.DateTime,
+        -- | The UTC time when the Android SafetyNet response was generated
+        generatedtime :: HG.DateTime
+      }
   | -- | The integrity check failed based on the required integrity from the
     -- format
-    VerificationErrorFailedIntegrityCheck Integrity
+    IntegrityCheckFailed Integrity
   deriving (Show, Exception)
 
 androidHostName :: VerificationHostName
@@ -215,7 +224,8 @@ instance M.AttestationStatementFormat Format where
     let signedData = rawData <> BA.convert (M.unClientDataHash clientDataHash)
     let hashedData = Hash.hashWith Hash.SHA256 signedData
     let encodedData = decodeUtf8 . Base64.encode $ BA.convert hashedData
-    unless (nonce response == encodedData) . failure $ VerificationErrorInvalidNonce
+    let responseNonce = nonce response
+    unless (responseNonce == encodedData) . failure $ NonceMismatch responseNonce encodedData
 
     -- 4. Verify that the SafetyNet response actually came from the SafetyNet service by following the steps in the
     -- SafetyNet online documentation.
@@ -236,15 +246,17 @@ instance M.AttestationStatementFormat Format where
     -- NOTE: For WebAuthn, we need not care about the package name or the app's signing certificate. The Nonce as
     -- has already been dealt with.
     let generatedTime = HG.timeConvert $ timestampMs response
-    when ((generatedTime `HG.timeAdd` driftBackwardsTolerance) < now) $ failure $ VerificationErrorResponseTooOld now generatedTime
-    when (generatedTime > (now `HG.timeAdd` driftForwardsTolerance)) $ failure $ VerificationErrorResponseInFuture now generatedTime
+    let lowerBound = now `HG.timeAdd` negate (HG.toSeconds driftBackwardsTolerance)
+    let upperBound = now `HG.timeAdd` driftForwardsTolerance
+    when (generatedTime < lowerBound) $ failure $ ResponseTimeInvalid lowerBound upperBound generatedTime
+    when (generatedTime > upperBound) $ failure $ ResponseTimeInvalid lowerBound upperBound generatedTime
 
     let integrity = case (basicIntegrity response, ctsProfileMatch response) of
           (_, True) -> CTSProfileIntegrity
           (True, False) -> BasicIntegrity
           (False, False) -> NoIntegrity
     unless (integrity >= requiredIntegrity) $
-      failure $ VerificationErrorFailedIntegrityCheck integrity
+      failure $ IntegrityCheckFailed integrity
 
     -- 5. If successful, return implementation-specific values representing attestation type Basic and attestation trust
     -- path x5c.